@@ -12,48 +12,52 @@
     "test-skip": "ng test --watch false",
     "lint": "run-p -snc lint:*",
     "lint:ng": "ng lint",
-    "lint:renderer": "tslint --project src/tsconfig.renderer.json",
-    "lint:style": "stylelint \"src/**/*.scss\" --syntax scss",
+    "lint:style": "stylelint \"src/**/*.scss\"",
     "pree2e": "webdriver-manager update --standalone false --gecko false --quiet",
     "e2e": "ng e2e"
   },
   "private": true,
   "license": "See in LICENSE",
   "devDependencies": {
-    "@angular/animations": "12.2.3",
-    "@angular-builders/custom-webpack": "12.1.1",
-    "@angular/cdk": "12.2.3",
-    "@angular/cli": "12.2.3",
-    "@angular/common": "12.2.3",
-    "@angular/compiler": "12.2.3",
-    "@angular/compiler-cli": "12.2.3",
-    "@angular/core": "12.2.3",
-    "@angular/localize": "12.2.3",
-    "@angular-devkit/build-angular": "12.2.3",
-    "@angular-devkit/build-optimizer": "0.1202.3",
-    "@angular-devkit/core": "12.2.3",
-    "@angular/forms": "12.2.3",
-    "@angular/language-service": "12.2.3",
-    "@angular/platform-browser": "12.2.3",
-    "@angular/platform-browser-dynamic": "12.2.3",
-    "@angular/router": "12.2.3",
+    "@angular-builders/custom-webpack": "13.0.0",
+    "@angular-devkit/build-angular": "13.0.3",
+    "@angular-devkit/build-optimizer": "0.1300.3",
+    "@angular-devkit/core": "13.0.3",
+    "@angular-eslint/builder": "13.0.1",
+    "@angular-eslint/eslint-plugin": "13.0.1",
+    "@angular-eslint/eslint-plugin-template": "13.0.1",
+    "@angular-eslint/schematics": "13.0.1",
+    "@angular-eslint/template-parser": "13.0.1",
+    "@angular/animations": "13.0.2",
+    "@angular/cdk": "13.0.2",
+    "@angular/cli": "13.0.3",
+    "@angular/common": "13.0.2",
+    "@angular/compiler": "13.0.2",
+    "@angular/compiler-cli": "13.0.2",
+    "@angular/core": "13.0.2",
+    "@angular/forms": "13.0.2",
+    "@angular/language-service": "13.0.2",
+    "@angular/localize": "13.0.2",
+    "@angular/platform-browser": "13.0.2",
+    "@angular/platform-browser-dynamic": "13.0.2",
+    "@angular/router": "13.0.2",
+    "@fortawesome/angular-fontawesome": "0.7.0",
+    "@fortawesome/fontawesome-svg-core": "1.2.30",
+    "@fortawesome/free-regular-svg-icons": "5.14.0",
+    "@fortawesome/free-solid-svg-icons": "5.14.0",
     "@ng-bootstrap/ng-bootstrap": "10.0.0",
-    "@ngtools/webpack": "12.2.3",
+    "@ng-select/ng-select": "8.1.1",
     "@ngrx/effects": "10.0.0",
     "@ngrx/router-store": "10.0.0",
     "@ngrx/store": "10.0.0",
     "@ngrx/store-devtools": "10.0.0",
-    "@ng-select/ng-select": "^7.2.0",
-    "ngx-contextmenu": "6.0.0",
+    "@ngtools/webpack": "13.0.3",
     "angular-confirmation-popover": "6.0.0",
     "angular-notifier": "9.1.0",
     "bootstrap": "4.6.0",
     "codelyzer": "6.0.0",
     "dragula": "3.7.2",
-    "@fortawesome/angular-fontawesome": "0.7.0",
-    "@fortawesome/fontawesome-svg-core": "1.2.30",
-    "@fortawesome/free-solid-svg-icons": "5.14.0",
-    "@fortawesome/free-regular-svg-icons": "5.14.0",
+    "eslint": "^8.2.0",
     "jasmine-core": "3.6.0",
     "jasmine-spec-reporter": "5.0.2",
     "karma": "6.3.4",
@@ -63,13 +67,14 @@
     "karma-jasmine-html-reporter": "1.7.0",
     "ng2-dragula": "2.1.1",
     "ng2-nouislider": "1.8.2",
-    "ngx-clipboard": "14.0.1",
-    "ngx-select-ex": "6.1.0",
     "ngrx-store-freeze": "0.2.4",
+    "ngx-clipboard": "14.0.1",
+    "ngx-contextmenu": "6.0.0",
+    "ngx-select-ex": "6.1.0",
     "nouislider": "14.6.0",
     "protractor": "7.0.0",
     "rxjs": "6.6.0",
-    "semver": "7.3.4",
+    "semver": "7.3.5",
     "ts-keycode-enum": "^1.0.6",
     "uhk-common": "1.0.0",
     "xml-loader": "1.2.1",
