@@ -1,4 +1,8 @@
-const settings = require('electron-settings');
-const { appTheme = 'system' } = JSON.parse(settings.get('application-settings') || '{}');
+const { ipcRenderer } =  require('electron')
 
-window.UHK_THEME = appTheme;
+ipcRenderer
+    .invoke('app-get-config', 'application-settings')
+    .then(appSettings => {
+        const { appTheme = 'system' } = JSON.parse(appSettings || '{}');
+        window.UHK_THEME = appTheme;
+    })
