@@ -2,8 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
     selector: 'main-page',
-    templateUrl: './main.page.html',
-    styles: [':host{height:100%; width:100%}']
+    templateUrl: './main.page.html'
 })
 export class MainPage {
 
